backend/requirements.txt
fastapi==0.127.0
mangum==0.19.0
uvicorn[standard]==0.40.0
boto3==1.42.16
pyjwt==2.10.1